services/users/requirements.txt
anyio==4.2.0
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.2.0
django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-heroku==0.3.1
django-select2==8.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
packaging==24.1
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
wheel==0.42.0
whitenoise==6.7.0
